requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
fastapi~=0.111.0